networking: Rename the `create` module.
