+++ crates/subspace-networking/examples/requests.rs
@@ -36,7 +36,7 @@ async fn main() {
         )],
         ..Config::default()
     };
-    let (node_1, mut node_runner_1) = subspace_networking::create(config_1).unwrap();
+    let (node_1, mut node_runner_1) = subspace_networking::compose(config_1).unwrap();
 
     println!("Node 1 ID is {}", node_1.id());
 
@@ -72,7 +72,7 @@ async fn main() {
         ..Config::default()
     };
 
-    let (node_2, mut node_runner_2) = subspace_networking::create(config_2).unwrap();
+    let (node_2, mut node_runner_2) = subspace_networking::compose(config_2).unwrap();
 
     println!("Node 2 ID is {}", node_2.id());
